Fix empty filter bug and refactor API
There was a bug when doing nothing with a video, just loading and
rendering it. In that case there would be no filters but the filter
option was still given to the command line, resulting in ffmpeg failing
with "unknown filter" error.

The functions Render and FFMPEG (now called CommandLine) were
refactored because they did the same thing: build the command line.