load: Support multiple streams
With some videos, there seems to be an array of streams,
defaulting to the first one may give unexpected results.

A `.mov` object, for example, in my case was returning the following:
```
desc.Streams=[{0 0 {<nil>}} {1280 720 {<nil>}}]
```

With this change, we return the correct size by ignoring streams that
have a width and a height of 0x0.